third_party/blink/web_tests/platform/win/external/wpt/editing/run/underline_1001-2000-expected.txt

This is a testharness.js-based test.
Found 176 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<strike>[bar]</strike>baz" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<strike>[bar]</strike>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[bar]baz</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<strike>foo<span style=\\"text-decoration:underline\\">bar</span>baz</strike>" but got "<strike>foo<span style=\\"text-decoration-line:underline\\">bar</span>baz</strike>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[bar]baz</strike>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[bar]baz</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<strike>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</strike>" but got "<strike>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</strike>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\">ar]ba</span>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<strike>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</strike>" but got "<strike>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</strike>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<strike>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</strike>" but got "<strike>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</strike>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<strike>foo<span style=\\"text-decoration:underline\\">b</span><i><span style=\\"text-decoration:underline\\">ar</span>ba</i>z</strike>" but got "<strike>foo<span style=\\"text-decoration-line:underline\\">b</span><i><span style=\\"text-decoration-line:underline\\">ar</span>ba</i>z</strike>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<strike>foo[b<i>ar]ba</i>z</strike>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<ins>[bar]</ins>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<ins>[bar]</ins>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[bar]baz</ins>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[bar]baz</ins>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<span style=\\"color:blue\\">ar]ba</span>z</ins>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</ins>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</ins>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "<ins>foo[b<i>ar]ba</i>z</ins>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<del>[bar]</del>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<del style=\\"text-decoration:underline line-through\\">bar</del>baz" but got "foo<del style=\\"text-decoration-line:underline\\">bar</del>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<del>[bar]</del>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<del>[bar]</del>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<del>[bar]</del>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<del>[bar]</del>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[bar]baz</del>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<del>foo<span style=\\"text-decoration:underline\\">bar</span>baz</del>" but got "<del>foo<span style=\\"text-decoration-line:underline\\">bar</span>baz</del>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[bar]baz</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[bar]baz</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[bar]baz</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[bar]baz</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<del>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</del>" but got "<del>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</del>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\">ar]ba</span>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<del>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</del>" but got "<del>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"color:rgb(0, 0, 255)\\" id=\\"foo\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</del>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<span style=\\"color:blue\\" id=foo>ar]ba</span>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<del>foo<span style=\\"text-decoration:underline\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration:underline\\">ar</span>ba</span>z</del>" but got "<del>foo<span style=\\"text-decoration-line:underline\\">b</span><span style=\\"font-size:3em\\"><span style=\\"text-decoration-line:underline\\">ar</span>ba</span>z</del>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<span style=\\"font-size:3em\\">ar]ba</span>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<i>ar]ba</i>z</del>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<del>foo<span style=\\"text-decoration:underline\\">b</span><i><span style=\\"text-decoration:underline\\">ar</span>ba</i>z</del>" but got "<del>foo<span style=\\"text-decoration-line:underline\\">b</span><i><span style=\\"text-decoration-line:underline\\">ar</span>ba</i>z</del>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<i>ar]ba</i>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<del>foo[b<i>ar]ba</i>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<i>ar]ba</i>z</del>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<del>foo[b<i>ar]ba</i>z</del>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\">bar</span>baz" but got "foo<strike>bar</strike>baz"
[FAIL] [["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\"><span style=\\"text-decoration:underline\\">b</span>a<span style=\\"text-decoration:underline\\">r</span></span>baz" but got "foo<span style=\\"text-decoration:underline line-through\\">b</span><span style=\\"text-decoration-line:line-through\\">a</span><span style=\\"text-decoration:underline line-through\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:line-through\\"><u>b</u>a<u>r</u></span>baz" but got "foo<span style=\\"text-decoration:underline line-through\\">b</span><strike>a</strike><span style=\\"text-decoration:underline line-through\\">r</span>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span style=\\"text-decoration: underline line-through\\">b[a]r</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">[bar]</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\">b</span>a<span style=\\"text-decoration:underline\\">r</span>baz" but got "foo<s style=\\"text-decoration:underline\\">b</s>a<s style=\\"text-decoration:underline\\">r</s>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>b</u>a<u>r</u>baz" but got "foo<s style=\\"text-decoration:underline\\">b</s>a<s style=\\"text-decoration:underline\\">r</s>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: underline\\">b[a]r</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\"><u style=\\"text-decoration:line-through\\">bar</u></span>baz" but got "foo<span style=\\"text-decoration-line:underline line-through\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u><u style=\\"text-decoration:line-through\\">bar</u></u>baz" but got "foo<u><strike>bar</strike></u>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">[bar]</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u style=\\"text-decoration:line-through\\">b<span style=\\"text-decoration:underline\\">a</span>r</u>baz" but got "foo<u style=\\"text-decoration:line-through\\">b</u><span style=\\"text-decoration-line:underline line-through\\">a</span><u style=\\"text-decoration:line-through\\">r</u>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u style=\\"text-decoration:line-through\\">b<u>a</u>r</u>baz" but got "foo<u style=\\"text-decoration:line-through\\">b</u><u><strike>a</strike></u><u style=\\"text-decoration:line-through\\">r</u>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: line-through\\">b[a]r</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline overline\\">bar</span>baz" but got "foo<span style=\\"text-decoration-line:underline overline\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u><s style=\\"text-decoration:overline\\">bar</s></u>baz" but got "foo<span style=\\"text-decoration-line:overline\\"><u>bar</u></span>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">[bar]</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s style=\\"text-decoration:overline\\">b<span style=\\"text-decoration:underline\\">a</span>r</s>baz" but got "foo<s style=\\"text-decoration:overline\\">b</s><span style=\\"text-decoration-line:underline overline\\">a</span><s style=\\"text-decoration:overline\\">r</s>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<s style=\\"text-decoration:overline\\">b<u>a</u>r</s>baz" but got "foo<s style=\\"text-decoration:overline\\">b</s><span style=\\"text-decoration-line:overline\\"><u>a</u></span><s style=\\"text-decoration:overline\\">r</s>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<s style=\\"text-decoration: overline\\">b[a]r</s>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\"><u style=\\"text-decoration:overline\\">bar</u></span>baz" but got "foo<span style=\\"text-decoration-line:underline overline\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u><u style=\\"text-decoration:overline\\">bar</u></u>baz" but got "foo<span style=\\"text-decoration-line:overline\\"><u>bar</u></span>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">[bar]</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u style=\\"text-decoration:overline\\">b<span style=\\"text-decoration:underline\\">a</span>r</u>baz" but got "foo<u style=\\"text-decoration:overline\\">b</u><span style=\\"text-decoration-line:underline overline\\">a</span><u style=\\"text-decoration:overline\\">r</u>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u style=\\"text-decoration:overline\\">b<u>a</u>r</u>baz" but got "foo<u style=\\"text-decoration:overline\\">b</u><span style=\\"text-decoration-line:overline\\"><u>a</u></span><u style=\\"text-decoration:overline\\">r</u>baz"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u style=\\"text-decoration: overline\\">b[a]r</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:overline\\">foo<span style=\\"text-decoration:underline\\">bar</span>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:overline\\">foo</span><span style=\\"text-decoration-line:underline overline\\">bar</span><span style=\\"text-decoration-line:overline\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p style=\\"text-decoration:overline\\">foo<u>bar</u>baz</p>" but got "<p style=\\"\\"><span style=\\"text-decoration-line:overline\\">foo<u>bar</u>baz</span></p>"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "<p style=\\"text-decoration: overline\\">foo[bar]baz</p>" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline\\">b[a]r</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span class=\\"line-through\\" style=\\"text-decoration:underline\\">bar</span>baz" but got "foo<span class=\\"line-through\\" style=\\"text-decoration-line:underline\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span class=\\"line-through\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span class=\\"line-through\\">b<span style=\\"text-decoration:underline\\">a</span>r</span>baz" but got "foo<span class=\\"line-through\\">b<span style=\\"text-decoration-line:underline\\">a</span>r</span>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<span class=\\"line-through\\">b[a]r</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline-and-line-through\\">[bar]</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline-and-line-through\\">[bar]</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline-and-line-through\\">b[a]r</span>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<span class=\\"underline-and-line-through\\">b[a]r</span>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "fo[o<u>b]ar</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "fo[o<u>b]ar</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["underline",""]] "foo<u>ba[r</u>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<u>barb</u>az" but got "foo<u>ba</u>rbaz"
[FAIL] [["underline",""]] "foo<u>ba[r</u>b]az" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["underline",""]] "foo<u>ba[r</u>b]az" queryCommandState("underline") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["underline",""]] "foo<u>ba[r</u>b]az" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "fo[o<u>bar</u>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"text-decoration:underline\\">obarb</span>az" but got "fo<span style=\\"text-decoration-line:underline\\">obarb</span>az"
[FAIL] [["stylewithcss","true"],["underline",""]] "fo[o<u>bar</u>b]az" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "fo[o<u>bar</u>b]az" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "fo[o<u>bar</u>b]az" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "fo[o<u>bar</u>b]az" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[<u>b]ar</u>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foob<span style=\\"text-decoration:underline\\">ar</span>baz" but got "foob<u>ar</u>baz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[<u>b]ar</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[<u>b]ar</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[<u>b]ar</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[<u>b]ar</u>baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u>ba[r</u>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"text-decoration:underline\\">ba</span>rbaz" but got "foo<u>ba</u>rbaz"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u>ba[r</u>]baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u>ba[r</u>]baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u>ba[r</u>]baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo<u>ba[r</u>]baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[<u>bar</u>]baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo[<u>bar</u>]baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[<u>bar</u>]baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["underline",""]] "foo[<u>bar</u>]baz" queryCommandValue("underline") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["underline",""]] "foo<u>[bar]</u>baz" queryCommandValue("underline") before
  assert_equals: Wrong result returned expected "" but got "true"
Harness: the test ran to completion.
